+++ scripts/lib/paths.js
@@ -4,10 +4,13 @@ const IG_ROOT = path.resolve(__dirname, '..', '..');
 const SOURCE_DATA_DIR = path.join(IG_ROOT, 'data');
 const FSH_DIR = path.join(IG_ROOT, 'input', 'fsh');
 const DATA_DIR = path.join(IG_ROOT, 'input', 'data');
+const STATIC_ASSETS_DIR = path.join(IG_ROOT, 'input', 'images');
+const GENERATED_STATIC_ASSETS_DIR = path.join(STATIC_ASSETS_DIR, 'generated');
 const GENERATED_DATA_DIR = path.join(DATA_DIR, 'generated');
 const GENERATED_DIR = path.join(FSH_DIR, 'generated');
 const GENERATED_SEARCH_PARAMETER_DIR = path.join(GENERATED_DIR, 'search-parameters');
 const GENERATED_FLAGS_PATH = path.join(GENERATED_DIR, 'USCDIQualityFlags.fsh');
+const USCDI_QUALITY_CSV_PATH = path.join(GENERATED_STATIC_ASSETS_DIR, 'uscdi-quality-data-elements.csv');
 const USCDI_QUALITY_DATA_PATH = path.join(SOURCE_DATA_DIR, 'uscdi_plus_quality.json');
 const REST_DATA_PATH = path.join(SOURCE_DATA_DIR, 'rest.json');
 
@@ -21,10 +24,13 @@ module.exports = {
   SOURCE_DATA_DIR,
   FSH_DIR,
   DATA_DIR,
+  STATIC_ASSETS_DIR,
+  GENERATED_STATIC_ASSETS_DIR,
   GENERATED_DATA_DIR,
   GENERATED_DIR,
   GENERATED_SEARCH_PARAMETER_DIR,
   GENERATED_FLAGS_PATH,
+  USCDI_QUALITY_CSV_PATH,
   USCDI_QUALITY_DATA_PATH,
   REST_DATA_PATH,
   USCDI_QUALITY_RULESET_PREFIX,